Importar y Exportar - Alcance 3: 5 - 9

+++ src/main/java/com/towers/huellacarbonobackend/dto/GeneracionResiduosDetalleDto.java
@@ -3,12 +3,12 @@ package com.towers.huellacarbonobackend.dto;
 public record GeneracionResiduosDetalleDto(
         Long id,
         Integer anio,
-        Integer productosMadera,
-        Integer productosPapel,
-        Integer residuos,
-        Integer textiles,
-        Integer jardines,
-        Integer paniales,
-        Integer otros
+        Double productosMadera,
+        Double productosPapel,
+        Double residuos,
+        Double textiles,
+        Double jardines,
+        Double paniales,
+        Double otros
 ) {
 }
